Update Modeicons, Integrate IrisSessionRowView & NewIrisSessionButton into IrisSessionListView
